+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### Bugfixes
   - fetch_metadata.py has more robust handling of a bad cached file (#168)
+  - Improved error message when a DELETE fails with 404 (#167)
 
 ## v2.12.0 (2022-12-08)
 #### New features